feat: 메인 컨트롤러 추가
